Sumaira

+++ src/main/java/net/atos/projetFinal/service/impl/ServiceAdresse.java
@@ -60,7 +60,7 @@ public class ServiceAdresse implements IAdresseService {
 	 */
 	@Override
 	@Transactional
-	public void supprimerAdresse(int idAdresse) {
+	public void supprimerAdresse(Long idAdresse) {
 		dao.deleteById(idAdresse);
 
 	}
